login front

+++ backend/controllers/wallpaperController.js
@@ -4,7 +4,23 @@ import request from 'request';
 
 // Get all Wallpapers
 const getWallpapers = async (req, res) => {
-    const wallpapers = await Wallpaper.find({}).sort({ createdAt: 1 });
+    const wallpapers = await Wallpaper.aggregate([
+        { $lookup : {
+            from: 'users',
+            localField: 'artistId',
+            foreignField: '_id',
+            as : "user"
+        } },
+        // Extraire la valeur de "pseudo" du tableau "user" avec $arrayElemAt
+        { $addFields: { "pseudo": { $arrayElemAt: ["$user.pseudo", 0] } } },
+
+        // Supprimer le champ "user" avec $project
+        { $project: { "user": 0 } },
+
+        // sort by latest
+        { $sort: { "createdAt": -1 } }
+
+    ])
 
     res.status(200).json(wallpapers);
 };
@@ -17,13 +33,29 @@ const getWallpaper = async (req, res) => {
         return res.status(404).json({ error: 'No such wallpaper'});
     }
     
-    const wallpaper = await Wallpaper.findById(id);
-    
-    if (!wallpaper) {
-        return res.status(400).json({ error: 'No such wallpaper'});
+    const objId = new mongoose.Types.ObjectId(id);
+    console.log(id)
+    const wallpapers = await Wallpaper.aggregate([
+        { $match: { _id: objId } },
+        { $lookup : {
+            from: 'users',
+            localField: 'artistId',
+            foreignField: '_id',
+            as : "user"
+        } },
+        // Extraire la valeur de "pseudo" du tableau "user" avec $arrayElemAt
+        { $addFields: { "pseudo": { $arrayElemAt: ["$user.pseudo", 0] } } },
+
+        // Supprimer le champ "user" avec $project
+        { $project: { "user": 0 } }
+
+    ])
+
+    if (wallpapers.length === 0){
+        return res.status(404).json({ error: 'No such wallpaper'});
     }
     
-    res.status(200).json(wallpaper);
+    res.status(200).json(wallpapers[0]);
 };
 
 // Get X Wallpapers
@@ -39,8 +71,19 @@ const getXWallpapers = async (req, res) => {
     const wallpapers = await Wallpaper.aggregate([
         { $match: { _id: { $nin : idArray } } },
         { $sample: { size: x } },
+        { $lookup : {
+            from: 'users',
+            localField: 'artistId',
+            foreignField: '_id',
+            as : "user"
+        } },
+        // Extraire la valeur de "pseudo" du tableau "user" avec $arrayElemAt
+        { $addFields: { "pseudo": { $arrayElemAt: ["$user.pseudo", 0] } } },
+
+        // Supprimer le champ "user" avec $project
+        { $project: { "user": 0 } }
+
     ])
-    
 
     res.status(200).json(wallpapers);
 };
@@ -67,10 +110,12 @@ const createWallpaper = async (req, res) => {
 
     // add doc to db
     try {
-        const wallpaper = await Wallpaper.create({ imageLink, artistId: user.pseudo, title, titleColor });
+        console.log(user._id)
+        const wallpaper = await Wallpaper.create({ imageLink, artistId: user._id, title, titleColor });
         res.status(200).json(wallpaper);
     }
     catch (err) {
+        console.log(err)
         res.status(400).json({error: err.message, emptyFields});
     }
 }
@@ -78,23 +123,30 @@ const createWallpaper = async (req, res) => {
 // DELETE a workout
 const deleteWallpaper = async (req, res) => {
     const { id } = req.params;
+    const user = req.user;
 
     if (!mongoose.Types.ObjectId.isValid(id)) {
         return res.status(404).json({ error: "No such wallpaper" });
     }
 
-    const wallpaper = await Wallpaper.findOneAndDelete({ _id: id });
+    const wallpaper = await Wallpaper.findOne({ _id: id });
 
     if (!wallpaper) {
         return res.status(400).json({ error: "No such wallpaper" });
+    }else {
+        if (user._id === wallpaper.artistId || user.role === 'admin') {
+            await Wallpaper.findOneAndDelete({ _id: id });
+            res.status(200).json(wallpaper);
+        } else {
+            res.status(401).json({ error: "request is not authorized" });
+        }
     }
-
-    res.status(200).json(wallpaper);
 };
 
 // UPDATE a Wallpaper
 const updateWallpaper = async (req, res) => {
     const { id } = req.params;
+    const user = req.user;
 
     let emptyFields = [];
 
@@ -102,15 +154,20 @@ const updateWallpaper = async (req, res) => {
         return res.status(404).json({ error: "No such wallpaper", emptyFields });
     }
 
-    const wallpaper = await Wallpaper.findOneAndUpdate({ _id: id }, {
-        ...req.body
-    })
+    const wallpaper = await Wallpaper.findOne({ _id: id });
 
     if (!wallpaper) {
         return res.status(400).json({ error: "No such wallpaper", emptyFields });
+    }else{
+        if (user._id === wallpaper.artistId || user.role === 'admin') {
+            await Wallpaper.findOneAndUpdate({ _id: id }, {
+                ...req.body
+            })
+            res.status(200).json(wallpaper);
+        }else{
+            res.status(401).json({ error: "request is not authorized" });
+        }
     }
-
-    res.status(200).json(wallpaper);
 };
 
 const downloadWallpaper = async (req, res) => {
